declutter model files
