+++ mambular/models/__init__.py
@@ -1,21 +1,32 @@
-from .mambular_classifier import MambularClassifier
-from .mambular_regressor import MambularRegressor
-from .fttransformer_regressor import FTTransformerRegressor
-from .fttransformer_classifier import FTTransformerClassifier
-from .mlp_classifier import MLPClassifier
-from .mlp_regressor import MLPRegressor
-from .tabtransformer_classifier import TabTransformerClassifier
-from .resnet_classifier import ResNetClassifier
-from .mambular_distributional import MambularLSS
+from .mambular import MambularClassifier, MambularRegressor, MambularLSS
+from .fttransformer import (
+    FTTransformerClassifier,
+    FTTransformerRegressor,
+    FTTransformerLSS,
+)
+from .mlp import MLPClassifier, MLPRegressor, MLPLSS
+from .tabtransformer import (
+    TabTransformerClassifier,
+    TabTransformerRegressor,
+    TabTransformerLSS,
+)
+from .resnet import ResNetClassifier, ResNetRegressor, ResNetLSS
+
 
 __all__ = [
     "MambularClassifier",
     "MambularRegressor",
-    "FTTransformerRegressor",
+    "MambularLSS",
     "FTTransformerClassifier",
+    "FTTransformerRegressor",
+    "FTTransformerLSS",
     "MLPClassifier",
     "MLPRegressor",
+    "MLPLSS",
     "TabTransformerClassifier",
+    "TabTransformerRegressor",
+    "TabTransformerLSS",
     "ResNetClassifier",
-    "MambularLSS",
+    "ResNetRegressor",
+    "ResNetLSS",
 ]
