+++ mambular/models/sklearn_base_classifier.py
@@ -125,6 +125,7 @@ class SklearnBaseClassifier(BaseEstimator):
         factor: float = 0.1,
         weight_decay: float = 1e-06,
         checkpoint_path="model_checkpoints",
+        dataloader_kwargs={},
         **trainer_kwargs
     ):
         """
@@ -164,6 +165,10 @@ class SklearnBaseClassifier(BaseEstimator):
             Factor by which the learning rate will be reduced.
         weight_decay : float, default=0.025
             Weight decay (L2 penalty) coefficient.
+        checkpoint_path : str, default="model_checkpoints"
+            Path where the checkpoints are being saved.
+        dataloader_kwargs: dict, default={}
+            The kwargs for the pytorch dataloader class.
         **trainer_kwargs : Additional keyword arguments for PyTorch Lightning's Trainer class.
 
 
@@ -191,6 +196,7 @@ class SklearnBaseClassifier(BaseEstimator):
             val_size=val_size,
             random_state=random_state,
             regression=False,
+            **dataloader_kwargs
         )
 
         self.data_module.preprocess_data(
