+++ mambular/models/sklearn_base_lss.py
@@ -149,6 +149,7 @@ class SklearnBaseLSS(BaseEstimator):
         weight_decay: float = 1e-06,
         checkpoint_path="model_checkpoints",
         distributional_kwargs=None,
+        dataloader_kwargs={},
         **trainer_kwargs
     ):
         """
@@ -190,6 +191,12 @@ class SklearnBaseLSS(BaseEstimator):
             Factor by which the learning rate will be reduced.
         weight_decay : float, default=0.025
             Weight decay (L2 penalty) coefficient.
+        distributional_kwargs : dict, default=None
+            any arguments taht are specific for a certain distribution.
+        checkpoint_path : str, default="model_checkpoints"
+            Path where the checkpoints are being saved.
+        dataloader_kwargs: dict, default={}
+            The kwargs for the pytorch dataloader class.
         **trainer_kwargs : Additional keyword arguments for PyTorch Lightning's Trainer class.
 
 
@@ -237,6 +244,7 @@ class SklearnBaseLSS(BaseEstimator):
             val_size=val_size,
             random_state=random_state,
             regression=True,
+            **dataloader_kwargs
         )
 
         self.data_module.preprocess_data(
